+++ CHANGES.md
@@ -1,5 +1,17 @@
+0.2.0.0
+=======
+
+-   Add support for mixing in package databases from stack snapshots.
+-   "stack" is now a reserved name when creating sandboxes.
+
+
+0.1.0.1
+=======
+
+-   Loosen version constraints.
+
+
 0.1.0.0
 =======
 
 -   Initial release.
-